Fix hold availability calculation with newer Koha versions
The holds availability calculation stopped working after the commit
"Bug 27069: Adapt uses of holdallowed" (1d9d05613b9) was introduced in
Koha. The return value of GetBranchItemRule() is now a string in newer
Koha versions. To keep the plugin working with older Koha versions we
keep the checks for integer return values as well.